requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.26.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
playwright==1.41.0
beautifulsoup4==4.12.3
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
aiosqlite==0.19.0
cryptography==42.0.2
email-validator==2.3.0
